fix: Whitespace in OO-dynamic constructs was silently dropped
Thanks go to Fred Orange who brought up this issue and provided
a test case to demonstrate and reproduce the issue.

Please note that this fix does not fully fix the issue as there are
still constellations that can cause this problem. This is e.g.
the case in JS if you place a comment between the dot and the
following identifier. A full fix here would require to have a
code parser watch the code and fixup offending instances to carry
over the correct styling/markup.

Signed-off-by: Benny Baumann <BenBE@geshi.org>