Change the semantics of _parseColor to allow partial arrays
This change allows _parseColor to use the defaults set in _parseCSS instead of providing its own default. It also allows to only change RGB or only change A without overriding the defaults.

Additionally, now it returns false in case of syntax error. _parseCSS reacts to that by not altering the default, so in a way returning false is equivalent to returning array().